backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.1
httpx==0.25.2
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
Flask==3.1.1
Flask-Cors==4.0.0
requests==2.32.3
